+++ source/Replays/ASAP/ReplayASAP.cpp
@@ -148,7 +148,7 @@ namespace rePlayer
         m_surround.Enable((settings && settings->overrideSurround) ? settings->surround : ms_surround);
     }
 
-    void ReplayASAP::SetSubsong(uint16_t subsongIndex)
+    void ReplayASAP::SetSubsong(uint32_t subsongIndex)
     {
         m_subsongIndex = subsongIndex;
         ResetPlayback();
