+++ source/Replays/FutureComposer/ReplayFutureComposer.cpp
@@ -130,7 +130,7 @@ namespace rePlayer
         m_surround.Enable((settings && settings->overrideSurround) ? settings->surround : ms_surround);
     }
 
-    void ReplayFutureComposer::SetSubsong(uint16_t subsongIndex)
+    void ReplayFutureComposer::SetSubsong(uint32_t subsongIndex)
     {
         m_subsongIndex = subsongIndex;
     }
